% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-analysis.r
\name{contact_analysis}
\alias{contact_analysis}
\title{Generate contact analysis data from samples.}
\usage{
contact_analysis(df, id, pos, type, type1, type2, value, max_dist = 50)
}
\arguments{
\item{df}{Dataframe containing sample data with required fields.}

\item{id}{ID code for line or drillhole (e.g., bhid).}

\item{pos}{Position of sample along the hole (i.e., depth; [to - from] / 2)}

\item{type}{Domain or zone code field (character or numeric field)}

\item{type1}{Type code 1 for analysis, will have negative distance in final data.}

\item{type2}{Type code 2 for analysis, will have positive distance in final data.}

\item{value}{A vaue field such as grade to analyse.}

\item{max_dist}{Maximum contact evaluation distance (default is 50 m)}
}
\value{
A list of two dataframes. Detail contains un-averaged raw data. Summary is averaged by distance.
}
\description{
Generate contact analysis data from samples.
}
\author{
Alex M. Trueman, 2018-03-21
}
