% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-analysis.r
\name{contact_plot}
\alias{contact_plot}
\title{Produce Basic Contact Analysis Plots.}
\usage{
contact_plot(data)
}
\arguments{
\item{data}{List of data frames created by \code{\link{contact_data}} with
'detail' and 'summary' contact data.}
}
\value{
A \code{\link{ggplot2}} plot object.
}
\description{
\code{contact_plot} produces a minimally formatted contact analysis plot
facetted by domain using \code{\link{ggplot2}}. The input data is created by
\code{\link{contact_data}}. Additonal formatting can be applied using
\code{\link{ggplot2}} functions (see example).
}
\examples{
library(ggplot2)
cdata <- contact_data(dholes, grade)
p <- contact_plot(cdata)
p +
 labs(x = "Distance (m)", y = "Mean grade (g/t)") +
 theme_bw()
}
