% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD.optim.R
\name{DD.optim}
\alias{DD.optim}
\title{Minimize the negative log likelihood to find the best-fitting density-dependent transmission function}
\usage{
DD.optim(FoI = FOI, N = Nref, gamma = truegamma)
}
\arguments{
\item{N}{The size of the mid-sized or reference population (Nref), which is used to scale the FOI to a truebeta for the DD function}

\item{gamma}{The true value of gamma, which is used as a mean for picking random starting values}

\item{FOI}{The true force of infection, which is used with nref to find a truebeta as a mean for picking random starting values}
}
\description{
'DD.optim' uses optim() to minimize the nll and fit the density-dependent transmission function
given random starting parameters centered around the true/known values for Beta and gamma.
}
\details{
You can use this function on it's own, but it is meant to go inside the optimization loop for all models
}
