% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD.optim.R
\name{FD.optim}
\alias{FD.optim}
\title{Minimize the negative log likelihood to find the best-fitting frequency-dependent transmission function}
\usage{
FD.optim(gamma, datasets, initial.inf, initial.sus, samp.sizes, pops,
  time.outs, time.samps)
}
\arguments{
\item{gamma}{The true value of gamma, which is used as a mean for picking random starting values}

\item{datasets}{Dataframe with dataset for each population as a column}

\item{initial.inf}{vector of initial number of infected individuals in each population}

\item{initial.sus}{vector of initial number of infected individuals in each population, which should be pops - initial.S}

\item{samp.sizes}{vector of sample sizes, in number of individuals, that were sampled to estimate prevalence at each time point}

\item{pops}{vector of total population sizes in each population, which should be initial.S + initial.I}

\item{time.outs}{a single value indicating how many time units (e.g., 21 days days) the ODE should be simulated for - should be as long as largest time.samp}

\item{time.samps}{vector of the times when the populations were sampled, such as 0, 7, 14, and 21 days}
}
\description{
'FD.optim' uses optim() to minimize the nll and fit the frequency-dependent transmission function
given random starting parameters centered around the true/known value for gamma. Given convergence issues,
the range of starting values for Beta is set between -2 and 0 (on the logscale).
}
\details{
You can use this function on it's own, but it is meant to go inside the optimization loop for all models
}
