% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-analysis.r
\name{interval_mid}
\alias{interval_mid}
\title{Extract the midpoint of an interval assigned to a data frame.}
\usage{
interval_mid(x, dp = 1)
}
\arguments{
\item{x}{A numeric vector or data frame column with intervals typically
applied using the base-R \code{cut} function.}

\item{dp}{A numeric scalar (default 1) defining the number of decimal places
for the base-R \code{round} function. May be needed to ensure that
mid-points are unique.}
}
\value{
A numeric vector containing the mid points of the intervals.
}
\description{
\code{interval_mid} Calculates the mid-point of an interval to be used as an
attractive label for plot axes when binning an axis. Intervals as defined by
functions such as base-R \code{\link{cut}}.
}
\author{
Alex M Trueman
}
\keyword{internal}
