% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NL.optim.R
\name{NL.optim}
\alias{NL.optim}
\title{Minimize the negative log likelihood to find the best-fitting nonlinear transmission function}
\usage{
NL.optim(beta = truebeta, gamma = truegamma)
}
\arguments{
\item{beta}{The true value of Beta, which is used as a mean for picking random starting values}

\item{gamma}{The true value of gamma, which is used as a mean for picking random starting values}
}
\description{
'NL.optim' uses optim() to minimize the nll and fit the nonlinear transmission function
given random starting parameters centered around the true/known values for Beta and gamma. The random starting
value for K is somewhere between ~0 and ~1 (on the logscale).
}
\details{
You can use this function on it's own, but it is meant to go inside the optimization loop for all models
}
