% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-analysis.r
\name{contact_analysis}
\alias{contact_analysis}
\title{Generate Contact Analysis Data from Drilhole Samples.}
\usage{
contact_analysis(x, id, pos, group, codes, value, max_dist = 50)
}
\arguments{
\item{x}{Data frame containing sample data with required fields identified in
the following arguments.}

\item{id}{Name of column in \code{x} containing numeric or character ID code
for each line or drillhole (e.g., bhid).}

\item{pos}{Name of column in \code{x} containing the numeric downhole depth
to the samples (i.e., depth; \code{[to - from] / 2})}

\item{group}{Name of column in \code{x} contaning the numeric or character
categorical code for the domains whose contacts are to be analysed.}

\item{codes}{Numeric or character vector with two values identifying the two
specific groups in \code{group} that will have their contacts analysed.}

\item{value}{Name of numeric column in \code{x} that will be analysed across
the contacts of \code{codes}.}

\item{max_dist}{Numeric, scalar maximum contact evaluation distance (default
is 50 m)}
}
\value{
A list of two data frames: 'detail' contains un-averaged raw data and
  'summary' is averaged by distance.
}
\description{
\code{contat_analysis} calculates the change in the mean of samples for a
specified value (e.g., grade) with distance eiter side of the contact between
two domins.
}
\examples{
dhdata <-  dplyr::mutate(dholes, depth = (to - from) / 2)
domains <- c(1100, 1001)
condata <- contact_analysis(dhdata, bhid, depth, domain, domains, grade, 20)
plot(condata$summary$distance, condata$summary$mean, type = "l")
}
\author{
Alex M. Trueman, 2018-03-21
}
