% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD.optim.R
\name{FD.optim}
\alias{FD.optim}
\title{Minimize the negative log likelihood to find the best-fitting frequency-dependent transmission function}
\usage{
FD.optim(randombeta = runif(1, -2, 0), gamma = truegamma)
}
\arguments{
\item{randombeta}{A random starting value for Beta on the logscale - works best between -2 and 0, but can be changed}

\item{gamma}{The true value of gamma, which is used as a mean for picking random starting values}
}
\description{
'FD.optim' uses optim() to minimize the nll and fit the frequency-dependent transmission function
given random starting parameters centered around the true/known value for gamma. Given convergence issues,
the range of starting values for Beta is set between -2 and 0 (on the logscale).
}
\details{
You can use this function on it's own, but it is meant to go inside the optimization loop for all models
}
